# alem-ekub-ui/src/app/services/models/page-response-payment-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PaymentResponse } from '../models/payment-response';
export interface PageResponsePaymentResponse {
  content?: Array<PaymentResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}