# alem-ekub-ui/src/app/services/models/payment-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PaymentRequest {
  amount: number;
  paymentMethod: string;
  remark: string;
  roundId: string;
  type: 'MEMBER_PAYMENT' | 'WINNING_PAYOUT';
  userId: string;
}