# alem-ekub-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Alem Equb",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habte",
            "email": "habtadugna@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "keycloak": []
        }
    ],
    "paths": {
...
        "/payments": {
            "get": {
                "tags": [
                    "payments"
                ],
                "operationId": "getPageOfPayments",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponsePaymentResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "payments"
                ],
                "operationId": "createPayment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/rounds/{ekub-id}": {
            "get": {
                "tags": [
                    "rounds"
                ],
                "operationId": "getEkubRounds",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RoundResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rounds/{ekub-id}/{version}/{round-no}": {
            "get": {
                "tags": [
                    "rounds"
                ],
                "operationId": "getRoundByEkubAndRoundNo",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "round-no",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoundResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rounds/user/pending-payments/{user-id}": {
            "get": {
                "tags": [
                    "rounds"
                ],
                "operationId": "getUserPendingPayments",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserPendingPaymentResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/payments/user/{user-id}": {
            "get": {
                "tags": [
                    "payments"
                ],
                "operationId": "getUserPayments",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/payments/user/round/{ekub-id}": {
            "get": {
                "tags": [
                    "payments"
                ],
                "operationId": "getUserRoundPayments",
                "parameters": [
                    {
                        "name": "ekub-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserRoundPaymentResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaymentRequest": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "MEMBER_PAYMENT",
                            "WINNING_PAYOUT"
                        ]
                    },
                    "userId": {
                        "type": "string"
                    },
                    "roundId": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "paymentMethod": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    }
                },
                "required": [
                    "amount",
                    "paymentMethod",
                    "remark",
                    "roundId",
                    "type",
                    "userId"
                ]
            },
...
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "profession": {
                        "type": "string"
                    },
                    "profilePicUrl": {
                        "type": "string"
                    },
                    "idCardImageUrl": {
                        "type": "string"
                    },
                    "guarantors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "ekubs": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ekubIds": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
...
            "RoundResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "ekubName": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roundNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "winner": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "paid": {
                        "type": "boolean"
                    },
                    "userGuarantees": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserGuaranteeResponse"
                        }
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UserGuaranteeResponse": {
                "type": "object",
                "properties": {
                    "guarantorId": {
                        "type": "string"
                    },
                    "guarantor": {
                        "type": "string"
                    },
                    "guaranteedId": {
                        "type": "string"
                    },
                    "guaranteed": {
                        "type": "string"
                    }
                }
            },
            "UserPendingPaymentResponse": {
                "type": "object",
                "properties": {
                    "equbName": {
                        "type": "string"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roundId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "roundNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "endDateTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "penalty": {
                        "type": "number"
                    },
                    "totalAmount": {
                        "type": "number"
                    }
                }
            },
            "PageResponsePaymentResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaymentResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PaymentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "ekubName": {
                        "type": "string"
                    },
                    "round": {
                        "$ref": "#/components/schemas/RoundResponse"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "paymentMethod": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "UserRoundPaymentResponse": {
                "type": "object",
                "properties": {
                    "row": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:9090/realms/alem-ekub/protocol/openid-connect/token",
                        "scopes": {}
                    }
                }
            }
        }
    }
}